Abstract out the labels and content in parser:
Instead of making a function which uses the parser object to determine
which labels to add and remove and create the report content, what we
will do is abstract it out in the class itself.

This will allow us to not need all the `require...` attributes and thus
no need to make a copy of them and cache them.

This will also allow the bot to be working on Python 3.6 and 3.7.