Update lock file and fix CI error
pytest requires `importlib_metadata` when Python version is < 3.8
Now, Pipfile.lock is updated on my local machine according to
Python version 3.8.6 and thus it doesn't think that `importlib_metadata`
is a dependency.

So, when CI is ran with Python 3.6 or 3.7, it fails as `pipenv` did not
install the dependency. The solution is to ignore the lock file and
install dependency with Pipfile.