Update visiting logic for the parser:
A class can contain doctest in the class-level docstring which
requires the visitor to skip doctest checking for all its method.
The updated way of visiting the nodes is to visit only the child
nodes and the child node will visit their child node if necessary.

This helps us in changing the ``skip_doctest`` value to ``True``
momentarily if the class contains doctest and reseting back to the
original value.